fix(component): fix #74 by minimizing nesting level of angular directive/components and divs
fix #74 by minimizing nesting level of angular components/directive and also divs. Key chalange is

that with Angular 2 componenets the selector element is not removed from DOM which means all the

formly config should have capablity to add css class to all the html tags which gets generated. This

makes configuration complicated. The other possible solution which was  chosen in this  PR  is to

minimize the div which gets generated which also ment that formly-fieldgroup was removed.

fix #74 and breaks formly.field.group is removed.